Remove ConfigMixin, turn it into a functions instead
