+++ src/megatron/hub/common/__init__.py
@@ -1,0 +1,4 @@
+from .config import ConfigProtocol
+from .model_provider_mixin import ModelProviderMixin
+
+__all__ = ["ConfigProtocol", "ModelProviderMixin"]
