+++ src/megatron/hub/models/gpt_provider.py
@@ -18,7 +18,7 @@ from megatron.core.transformer.module import MegatronModule
 from megatron.core.transformer.transformer_config import TransformerConfig
 from megatron.core.utils import get_te_version
 
-from megatron.hub.common.mixins.model_mixin import ModelProviderMixin
+from megatron.hub.common.model_provider_mixin import ModelProviderMixin
 from megatron.hub.core.utils import fusions
