+++ src/megatron/hub/models/t5_provider.py
@@ -21,7 +21,7 @@ from megatron.core.models.T5.t5_model import T5Model as MCoreT5Model
 from megatron.core.transformer.spec_utils import ModuleSpec
 from megatron.core.transformer.transformer_config import TransformerConfig
 
-from megatron.hub.common.mixins.model_mixin import ModelProviderMixin
+from megatron.hub.common.model_provider_mixin import ModelProviderMixin
 from megatron.hub.core.utils.import_utils import safe_import
 from megatron.hub.models.gpt_provider import get_vocab_size
